typhoeus: hacky way to trace also hydra calls
This adjust the typhoeus instrumentation to support also calls executed
by hydra (multiple requests in parallel). To be able to represent the
requests as segments all requests are assumented to run serial but with
original timing information. This means that the total time of the
typhoeus hydra call is (much) smaller than the sum of the child segments
(external http requests).